feat: version methods
Allows the user to provide a flag/config that will dictate how to calculate versions. Currently this is bound to solely

- Increment Major Version
- Increment Minor Version
- Increment Patch Version

If their was no previous version set to a default of "v0.1.0"

If the previous version was not semver, then keep the previous name. This handles cases like "Beta"/"Alpha"