refactor: remove edit/template that edit files
It's so much simpler to allow the user to hot reload the config rather than having them be able to edit the template and edit the preview itself. This comes with a downside where you can't change the branch/version/preview for a specific preview, but this can be resolved other ways that should be simpler and more interactive.